requirements.txt
arrow==1.2.2
azure-functions==1.10.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
oauthlib==3.2.0
python-dateutil==2.8.2
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.1
tweepy==4.7.0
urllib3==1.26.9
